things

+++ src/controllers/api/apiDropController.js
@@ -22,11 +22,22 @@ import provinces from "../../utils/staticDB/provinces.js";
 import { getMappedErrors } from "../../utils/helpers/getMappedErrors.js";
 import { HTTP_STATUS } from "../../utils/staticDB/httpStatusCodes.js";
 import { deleteSensitiveUserData } from "./apiUserController.js";
-import { uploadFilesToAWS } from "../../utils/helpers/awsHandler.js";
-import { insertFilesInDB } from "../../utils/helpers/filesHandler.js";
+import {
+  destroyFilesFromAWS,
+  getFilesFromAWS,
+  uploadFilesToAWS,
+} from "../../utils/helpers/awsHandler.js";
+import {
+  deleteFileInDb,
+  insertFilesInDB,
+} from "../../utils/helpers/filesHandler.js";
 import entityTypes from "../../utils/staticDB/entityTypes.js";
 import sections from "../../utils/staticDB/sections.js";
 import getFileType from "../../utils/helpers/getFileType.js";
+import {
+  productIncludeArray,
+  setProductKeysToReturn,
+} from "./apiProductController.js";
 
 const DROPS_FOLDER_NAME = "drops";
 // ENV
@@ -34,10 +45,14 @@ const DROPS_FOLDER_NAME = "drops";
 const controller = {
   getDrops: async (req, res) => {
     try {
-      let { drops_id } = req.query;
+      let { drops_id, includeProductImages } = req.query;
       drops_id = drops_id || undefined;
-
-      let dropsFromDB = await getDropsFromDB(drops_id);
+      includeProductImages = includeProductImages ? true : false;
+      let dropsFromDB = await getDropsFromDB({
+        id: drops_id,
+        withImages: true,
+        withProductImages: includeProductImages,
+      });
 
       // Mando la respuesta
       return res.status(HTTP_STATUS.OK.code).json({
@@ -47,7 +62,7 @@ const controller = {
           method: "GET",
         },
         ok: true,
-        drops: dropsFromDB,
+        data: dropsFromDB,
       });
     } catch (error) {
       console.log(`Falle en apiDropController.getDrops`);
@@ -83,40 +98,33 @@ const controller = {
           .json({ msg: HTTP_STATUS.INTERNAL_SERVER_ERROR.message });
 
       // Ahora agarro los productos que relaciono
-      let { products_id } = req.body;
-      let relationsToPush = products_id.map((prodID) => ({
+      let { productIDS } = req.body;
+      productIDS = JSON.parse(productIDS);
+      let relationsToPush = productIDS?.map((prodID) => ({
         products_id: prodID,
-        drop_id: createdDrop.id,
+        drops_id: createdDrop.id,
       }));
       relationsToPush.length &&
         (await db.Product_Drop.bulkCreate(relationsToPush));
       // Ahora, si cargo fotos...
       let { files } = req;
       let { filesFromArray } = req.body; //Esto es lo que me llega por body
+      filesFromArray = filesFromArray ? JSON.parse(filesFromArray) : [];
       if (files && files.length) {
-        files?.forEach((multerFile) => {
-          const fileFromFilesArrayFiltered = filesFromArray.find(
-            (arrFile) => arrFile.filename == multerFile.originalname
-          );
-          multerFile.file_types_id = getFileType(multerFile);
-          multerFile.file_roles_id = fileFromFilesArrayFiltered.file_roles_id;
-        });
-        const objectToUpload = {
+        const filesToInsertInDb = await handleDropFilesUpload({
           files,
-          folderName: DROPS_FOLDER_NAME,
-          sections_id: sections.DROP.id, //Drop
-        };
-        const filesToInsertInDb = await uploadFilesToAWS(objectToUpload);
+          filesFromBody: filesFromArray,
+        });
         if (!filesToInsertInDb) {
           return res.status(HTTP_STATUS.INTERNAL_SERVER_ERROR.code).json({
             ok: false,
-            msg: createFailed,
+            msg: systemMessages.dropMsg.createFailed,
           });
         }
         const isInsertingFilesSuccessful = await insertFilesInDB({
           files: filesToInsertInDb,
           entities_id: createdDrop.id,
-          entity_types_id: entityTypes.DROP
+          entity_types_id: entityTypes.DROP,
         });
         if (!isInsertingFilesSuccessful) {
           return res.status(HTTP_STATUS.INTERNAL_SERVER_ERROR.code).json({
@@ -162,10 +170,113 @@ const controller = {
           msg: systemMessages.formMsg.validationError,
         });
       }
+      let { id } = req.params;
+      if (!id)
+        return res.status(HTTP_STATUS.BAD_REQUEST.code).json({
+          ok: false,
+        });
+      const dbDrop = await getDropsFromDB({ id });
+      if (!dbDrop) {
+        return res.status(HTTP_STATUS.NOT_FOUND.code).json({
+          ok: false,
+          msg: "Drop no encontrado",
+        });
+      }
       let dropObjToDB = generateDropObject(req.body);
-
+      dropObjToDB.id = id;
       await updateDropFromDB(dropObjToDB, dropObjToDB.id);
+      console.log("🟢 Drop actualizado con éxito:", dropObjToDB);
+      // Ahora veo que productos tiene
+      let { productIDS } = req.body;
+      productIDS = JSON.parse(productIDS);
+      // Calcular qué productos agregar y cuáles eliminar
+      let { idsToAdd, idsToRemove } = getDropProductRelationLists({
+        productIDS,
+        dbDrop,
+      });
+      const productRelationsToAdd =
+        idsToAdd?.map((prodID) => ({
+          products_id: prodID,
+          drops_id: dbDrop.id,
+        })) || [];
+      productRelationsToAdd.length &&
+        (await db.Product_Drop.bulkCreate(productRelationsToAdd));
+      idsToRemove.length &&
+        (await db.Product_Drop.destroy({
+          where: {
+            products_id: idsToRemove,
+            drops_id: dbDrop.id,
+          },
+        }));
+      console.log("🟢 Relaciones con productos actualizadas con éxito:");
 
+      // Tema Imagenes
+      const dbFiles = dbDrop?.files || [];
+      let filesToKeep = req.body.current_images;
+      filesToKeep = filesToKeep ? JSON.parse(filesToKeep) : [];
+      // current_images
+      // [
+      // id: fileid
+      // filename: randomName
+      // main_image: 1,
+      // position: 1
+      //]
+      // filesFromArray
+      // [
+      // filename: filename
+      // main_image: 0,
+      //position: 2
+      // ]
+      // req.files
+      let filesToDelete;
+      if (filesToKeep && filesToKeep.length > 0) {
+        filesToDelete = dbFiles.filter(
+          (img) =>
+            !filesToKeep.map((img) => img.filename).includes(img.filename)
+        );
+      } else {
+        filesToDelete = dbFiles;
+      }
+      const filesDeletedOK = await handleDropFilesDestroy(filesToDelete);
+      if (!filesDeletedOK)
+        return res.status(HTTP_STATUS.INTERNAL_SERVER_ERROR.code).json({
+          ok: false,
+          msg: updateFailed,
+        });
+
+      let normalizedFilesToUpdateInDb = filesToKeep.map((file) => ({
+        ...file,
+      }));
+      let { files } = req;
+      let { filesFromArray } = req.body;
+      filesFromArray = filesFromArray ? JSON.parse(filesFromArray) : [];
+      if (files && files.length) {
+        let filesToInsertInDb = await handleDropFilesUpload({
+          files,
+          filesFromBody: filesFromArray,
+        });
+        if (!filesToInsertInDb) {
+          return res.status(HTTP_STATUS.INTERNAL_SERVER_ERROR.code).json({
+            ok: false,
+            msg: systemMessages.dropMsg.createFailed,
+          });
+        }
+        normalizedFilesToUpdateInDb = [
+          ...normalizedFilesToUpdateInDb,
+          ...filesToInsertInDb,
+        ];
+      }
+      const isInsertingFilesSuccessful = await insertFilesInDB({
+        files: normalizedFilesToUpdateInDb,
+        entities_id: dbDrop.id,
+        entity_types_id: entityTypes.DROP,
+      });
+      if (!isInsertingFilesSuccessful) {
+        return res.status(HTTP_STATUS.INTERNAL_SERVER_ERROR.code).json({
+          ok: false,
+          msg: systemMessages.dropMsg.createFailed,
+        });
+      }
       // Le  mando ok con el redirect al email verification view
       return res.status(HTTP_STATUS.OK.code).json({
         meta: {
@@ -184,11 +295,30 @@ const controller = {
   },
   destroyDrop: async (req, res) => {
     try {
-      let { id } = req.body;
+      let { id } = req.params;
+      if (!id)
+        return res.status(HTTP_STATUS.BAD_REQUEST.code).json({
+          ok: false,
+        });
+      const dbDrop = await getDropsFromDB({ id });
+      if (!dbDrop) {
+        return res.status(HTTP_STATUS.NOT_FOUND.code).json({
+          ok: false,
+          msg: "Drop no encontrado",
+        });
+      }
       // Lo borro de db
       let response = await destroyDropFromDB(id);
       if (!response)
         return res.status(HTTP_STATUS.INTERNAL_SERVER_ERROR.code).json();
+
+      //Ahora borro los files
+      let filesDeletedOK = await handleDropFilesDestroy(dbDrop.files);
+      if (!filesDeletedOK)
+        return res.status(HTTP_STATUS.INTERNAL_SERVER_ERROR.code).json({
+          ok: false,
+          msg: systemMessages.dropMsg.destroyFailed,
+        });
       return res.status(HTTP_STATUS.OK.code).json({
         meta: {
           status: HTTP_STATUS.OK.code,
@@ -209,7 +339,10 @@ const controller = {
 
 export default controller;
 
-let dropIncludeArray = ["products"];
+let dropIncludeArray = [
+  { association: "products", include: productIncludeArray },
+  "files",
+];
 
 export async function insertDropToDB(obj) {
   try {
@@ -259,59 +392,148 @@ export async function destroyDropFromDB(id) {
 
 export function generateDropObject(obj) {
   // objeto para armar la address
-  let { id, name, active, unique } = obj;
+  let { name, active, unique, launch_date } = obj;
 
   let dataToDB = {
-    id: id ? id : uuidv4(),
-    name: name.trim().toLowerCase(),
+    id: uuidv4(),
+    name: name.trim(),
     active: active || false,
     unique: unique || false,
+    launch_date: new Date(launch_date), // Convertir a objeto Date
   };
   return dataToDB;
 }
 
-export async function getDropsFromDB(id = undefined) {
+export async function getDropsFromDB({
+  id = undefined,
+  withImages = false,
+  withProductImages = false,
+}) {
   try {
-    let colorsToReturn, colorToReturn;
+    let dropsToReturn, dropToReturn;
     // Condición si id es un string
     if (typeof id === "string") {
-      colorToReturn = await db.Drop.findByPk(id, {
+      dropToReturn = await db.Drop.findByPk(id, {
         include: dropIncludeArray,
       });
-      if (!colorToReturn) return null;
-      colorToReturn = colorToReturn && getDeepCopy(colorToReturn);
-      setDropKeysToReturn(colorToReturn);
-      return colorToReturn;
+      if (!dropToReturn) return null;
+      dropToReturn = dropToReturn && getDeepCopy(dropToReturn);
+      await setDropKeysToReturn({
+        drop: dropToReturn,
+        withProductImages,
+        withImages,
+      });
+      return dropToReturn;
     }
 
     // Condición si id es un array
     else if (Array.isArray(id)) {
-      colorsToReturn = await db.Drop.findAll({
+      dropsToReturn = await db.Drop.findAll({
         where: {
           id: id, // id es un array, se hace un WHERE id IN (id)
         },
         include: dropIncludeArray,
       });
-      if (!colorsToReturn || !colorsToReturn.length) return null;
-      colorsToReturn = getDeepCopy(colorsToReturn);
+      if (!dropsToReturn || !dropsToReturn.length) return null;
+      dropsToReturn = getDeepCopy(dropsToReturn);
     }
     // Condición si id es undefined
     else {
-      colorsToReturn = await db.Drop.findAll({
+      dropsToReturn = await db.Drop.findAll({
         include: dropIncludeArray,
       });
-      if (!colorsToReturn || !colorsToReturn.length) return null;
-      colorsToReturn = getDeepCopy(colorsToReturn);
+      if (!dropsToReturn || !dropsToReturn.length) return null;
+      dropsToReturn = getDeepCopy(dropsToReturn);
     }
-    colorToReturn?.forEach((color) => setDropKeysToReturn(color));
-    return colorsToReturn;
+    for (let i = 0; i < dropsToReturn.length; i++) {
+      const drop = dropsToReturn[i];
+      await setDropKeysToReturn({ drop, withProductImages, withImages });
+    }
+    return dropsToReturn;
   } catch (error) {
-    console.log("Falle en getColorsFromDB");
+    console.log("Falle en getdropsFromDB");
     console.error(error);
     return null;
   }
 }
 
-function setDropKeysToReturn(drop) {
-  //   drop.name = capitalizeFirstLetterOfEachWord(drop.name, true);
+async function setDropKeysToReturn({
+  drop,
+  withImages = false,
+  withProductImages = false,
+}) {
+  drop.products.forEach(
+    async (dropProd) =>
+      await setProductKeysToReturn({
+        product: dropProd,
+        withImages: withProductImages,
+      })
+  );
+  if (withImages && drop.files?.length) {
+    await getFilesFromAWS({
+      folderName: DROPS_FOLDER_NAME,
+      files: drop.files,
+    });
+  }
+}
+
+async function handleDropFilesUpload({ files = [], filesFromBody = [] }) {
+  try {
+    if (!files.length || !filesFromBody.length) return false;
+    files?.forEach((multerFile) => {
+      const fileFromFilesArrayFiltered = filesFromBody.find(
+        (arrFile) => arrFile.filename == multerFile.originalname
+      );
+      multerFile.file_types_id = getFileType(multerFile);
+      multerFile.file_roles_id =
+        fileFromFilesArrayFiltered.file_roles_id || null;
+    });
+    const objectToUpload = {
+      files,
+      folderName: DROPS_FOLDER_NAME,
+      sections_id: sections.DROP.id, //Drop
+    };
+    const filesToInsertInDb = await uploadFilesToAWS(objectToUpload);
+    if (!filesToInsertInDb) {
+      return false;
+    }
+    return filesToInsertInDb;
+  } catch (error) {
+    console.log(error);
+
+    return false;
+  }
+}
+
+async function handleDropFilesDestroy(files = []) {
+  try {
+    if (!files.length) return true;
+    const objectToDestroyInAws = {
+      files,
+      folderName: DROPS_FOLDER_NAME,
+    };
+    const isDeletionInAwsSuccessful = await destroyFilesFromAWS(
+      objectToDestroyInAws
+    );
+    if (!isDeletionInAwsSuccessful) return false;
+    const idsToDestroyDB = files.map((img) => img.id);
+    await deleteFileInDb(idsToDestroyDB);
+    return true;
+  } catch (error) {
+    console.log(error);
+    return false;
+  }
+}
+
+function getDropProductRelationLists({ productIDS = [], dbDrop = {} }) {
+  // Obtener IDs de productos actualmente relacionados
+  const currentProductIds = dbDrop?.products.map((p) => p.id) || [];
+  // Calcular diferencias
+  const productsToRemove = currentProductIds.filter(
+    (id) => !productIDS.includes(id)
+  ); // Estaban antes, pero ya no están
+  const productsToAdd = productIDS.filter(
+    (id) => !currentProductIds.includes(id)
+  ); // No estaban antes, pero ahora sí
+  return { idsToAdd: productsToAdd, idsToRemove: productsToRemove };
 }
